Updated tests for package portal
