+++ portal/example_portal_test.go
@@ -9,7 +9,6 @@ const TestPhrase = "Hello Portal!"
 
 func ExamplePortal() {
 	p := portal.New("echo", TestPhrase)
-
 	p.WriteTo(os.Stdout)
 	// Output: Hello Portal!
 }
