+++ portal/portal.go
@@ -12,12 +12,14 @@ type Portal struct {
 	blueprint *exec.Cmd
 }
 
-// New creates a new Portal using exec.Command internally.
+// New creates a new, unprepared Portal.
+//
+// It uses exec.Command internally to generate the command.
 func New(name string, args ...string) *Portal {
 	return MakeFrom(exec.Command(name, args...))
 }
 
-// MakeFrom creates a Portal by wrapping exec.Cmd.
+// MakeFrom creates a new, unprepared Portal by wrapping exec.Cmd.
 //
 // If "cmd" is nil, MakeFrom will return nil.
 func MakeFrom(cmd *exec.Cmd) *Portal {
@@ -27,8 +29,18 @@ func MakeFrom(cmd *exec.Cmd) *Portal {
 	return &Portal{blueprint: cmd}
 }
 
-// Restore loads Portal's exec.Cmd instance from an internal reference,
+// Prepare loads Portal's exec.Cmd instance from an internal reference,
 // preparing it to perform a comamnd.
-func (p *Portal) Restore() {
+//
+// Prepare must be called before each execution of Portal's internal Cmd.
+func (p *Portal) Prepare() {
 	p.Cmd = *p.blueprint
+}
+
+// IsReady determines if the Portal has been "prepared" to be executed.
+//
+// If IsReady returns false, that means Portal's internal Cmd field is empty
+// and will fail if started.
+func (p *Portal) IsReady() bool {
+	return p.Cmd.Path != ""
 }
