+++ portal/portal_test.go
@@ -1,46 +1,30 @@
 package portal_test
 
 import (
-	"bytes"
 	"github.com/steven-xie/glip/portal"
+	"os/exec"
 	"testing"
 )
 
-func TestPortal_Read(t *testing.T) {
-	p := portal.New("echo", TestPhrase)
+// Global Portal instances to be used by the testing package.
+var (
+	catp  = portal.New("cat")
+	echop = portal.New("echo", TestPhrase)
+)
 
-	data := make([]byte, len(TestPhrase))
-	if _, err := p.Read(data); err != nil {
-		t.Fatal("Failed to read data from portal:", err)
+func TestPortal(t *testing.T) {
+	var (
+		echoCmd    = exec.Command("echo", TestPhrase)
+		echoPortal = portal.MakeFrom(echoCmd)
+	)
+
+	if echoPortal.Path != "" {
+		t.Fatalf("Expected new portal to begin with an empty Cmd field (with an "+
+			"empty Path), but found echoPortal.Cmd to be: %#v", echoPortal.Cmd)
 	}
 
-	outstr := string(data)
-	if outstr != TestPhrase {
-		t.Errorf("Expected an output of \"%s\", instead got: \"%s\"", TestPhrase,
-			outstr)
-	}
-}
-
-var catp = portal.New("cat")
-
-func TestPortal_ReadFrom(t *testing.T) {
-	buf := bytes.NewBufferString(TestPhrase)
-
-	n, err := catp.ReadFrom(buf)
-	if err != nil {
-		t.Fatal("Could not read from buffer to portal:", err)
-	}
-	if n == 0 {
-		t.Error("Portal unexpectedly read 0 bytes from buffer.")
-	}
-}
-
-func TestPortal_Write(t *testing.T) {
-	n, err := catp.Write([]byte(TestPhrase))
-	if err != nil {
-		t.Fatal("Could not write to portal:", err)
-	}
-	if n == 0 {
-		t.Error("Portal unexpectedly wrote no data.")
+	echoPortal.Prepare()
+	if !echoPortal.IsReady() {
+		t.Fatal("Expected prepared portal to be ready for execution.")
 	}
 }
